라우터를 컴포넌트로 정립하여 추가.
