+++ frontend/src/App.tsx
@@ -1,42 +1,8 @@
 import React, { useEffect, useState } from 'react'
-
 import Layout from './pages/Layout'
 import axios from 'axios'
 import database from './tempDatabase'
-import { Redirect, Route, Switch } from 'react-router-dom'
-
-import PageHome from './pages/Home'
-import PageJoin from './pages/Join'
-import PageLogin from './pages/Login'
-import PageCounsel from './pages/Counsel'
-import PageManage from './pages/Manage'
-import PageNotFound from './pages/NotFound'
-
-export type User = {
-  id: number
-  email: string
-  name: string
-  sv_number: string
-  phone_number: string
-  authority: number
-}
-
-export function parseAuthority(value: number | null | undefined) {
-  switch (value) {
-    case 1:
-      return '서비스관리자'
-    case 2:
-      return '부대관리자'
-    case 3:
-      return '상담관'
-    case 4:
-      return '지휘관'
-    case 5:
-      return '일반사용자'
-    default:
-      return '로그인 정보 없음'
-  }
-}
+import { Router, User } from './routes'
 
 export default function App() {
   const [user, setUser] = useState<User | null>(null)
@@ -75,22 +41,7 @@ export default function App() {
 
   return (
     <Layout user={user} changeUser={authenticateUser}>
-      <Switch>
-        <Route exact path="/" render={() => <PageHome user={user} />} />
-        {user ? (
-          <>
-            <Route path="/counsel" render={() => <PageCounsel user={user} />} />
-            <Route path="/manage" render={() => <PageManage user={user} />} />
-          </>
-        ) : (
-          <>
-            <Route path="/login" render={() => <PageLogin user={user} />} />
-            <Route path="/join" render={() => <PageJoin user={user} />} />
-          </>
-        )}
-
-        <Route render={() => <PageNotFound user={user} />} />
-      </Switch>
+      <Router user={user} />
     </Layout>
   )
 }
