+++ frontend/src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { User } from '../App'
+import { User } from '../routes'
 
 type HomeProps = {
   user: User | null
