+++ frontend/src/pages/NotFound.tsx
@@ -1,10 +1,7 @@
 import React from 'react'
 import { User } from '../App'
 
-type NotFoundProps = {
-  user: User | null
-}
 
-export default function NotFound({ user }: NotFoundProps) {
+export default function NotFound() {
   return <div>404 Not Found</div>
 }
